[CIAS30-3697] Move intervention elements directions selectors to global state reducer
